iss-907: added support for simple command after ajc compilation
