+++ gradle.properties
@@ -1,0 +1,1 @@
+org.gradle.java.home=c:/Program Files/Java/jdk1.7.0_79
