+++ hystrix-contrib/hystrix-javanica/src/ajcTest/resources/log4j.properties
@@ -1,0 +1,10 @@
+# Define the root logger with appender console
+log4j.rootLogger = ERROR, CONSOLE
+
+# Define the console appender
+log4j.appender.CONSOLE=org.apache.log4j.ConsoleAppender
+log4j.appender.CONSOLE.File=${log}/log.out
+
+# Define the layout for console appender
+log4j.appender.CONSOLE.layout=org.apache.log4j.PatternLayout
+log4j.appender.CONSOLE.layout.conversionPattern=%m%n
