+++ hystrix-contrib/hystrix-javanica/src/test/java/com/netflix/hystrix/contrib/javanica/test/common/domain/Profile.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.netflix.hystrix.contrib.javanica.test.spring.domain;
+package com.netflix.hystrix.contrib.javanica.test.common.domain;
 
 /**
  * Created by dmgcodevil on 1/9/2015.
